requirements.txt
Pillow
numpy
pandas
streamlit
tensorflow==2.15.0